requirements.txt
# Application Requirements
PySide6>=6.2.0
Pillow>=9.2.0
pynput>=1.7.6
PyYAML>=6.0